Complete tests for stores query
